fix(tests): resolve phpstan and phpcs violations
- Declare magento/module-configurable-product in composer.json require
  so PHPStan can resolve Configurable and ConfigurableResource types
- Add use stdClass import; remove inline fully-qualified reference
- Extract AddToCartExposed and QuoteItemWithProductId to separate files
  to satisfy PSR1 one-class-per-file rule
